+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Remove temporary code for data setup [#137](https://github.com/rokwire/gateway-building-block/issues/137)
+
 ## [2.14.0] - 2025-03-14
 ### Changed
 - Events summary APIs - part 1 [#132](https://github.com/rokwire/gateway-building-block/issues/132)