+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- Expose get legacy events and get summary events API [#135](https://github.com/rokwire/gateway-building-block/issues/135)
+- Expose get summary events API [#135](https://github.com/rokwire/gateway-building-block/issues/135)
 
 ### Changed
 - Remove temporary code for data setup [#137](https://github.com/rokwire/gateway-building-block/issues/137)